feat: move types to Sanity folder
