+++ src/pages/api/code-snippets.ts
@@ -1,10 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import { codeSnippetsQuery } from '@/queries/codeSnippets';
+import { CodeSnippet } from '@/sanityTypes/CodeSnippet';
 
 import { sanityClient } from '../../../sanity/lib/client';
-
-import { CodeSnippet } from '@/types/CodeSnippet';
 
 const codeSnippetsApi = async (req: NextApiRequest, res: NextApiResponse) => {
   const codeSnippets: CodeSnippet[] = await sanityClient.fetch(
