+++ src/pages/api/home.ts
@@ -1,10 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import { homeContentQuery } from '@/queries/homeContent';
+import { HomeContent } from '@/sanityTypes/HomeContent';
 
 import { sanityClient } from '../../../sanity/lib/client';
-
-import { HomeContent } from '@/types/HomeContent';
 
 const homeApi = async (req: NextApiRequest, res: NextApiResponse) => {
   const homeContent: HomeContent = await sanityClient.fetch(homeContentQuery);
