+++ src/pages/api/jobs.ts
@@ -1,10 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import { jobsQuery } from '@/queries/jobs';
+import { Job } from '@/sanityTypes/Job';
 
 import { sanityClient } from '../../../sanity/lib/client';
-
-import { Job } from '@/types/Job';
 
 const jobsApi = async (req: NextApiRequest, res: NextApiResponse) => {
   const jobs: Job[] = await sanityClient.fetch(jobsQuery);
