+++ src/pages/api/random-facts.ts
@@ -6,10 +6,9 @@ import {
   falseRandomFactsQuery,
   selectedTrueRandomFactsQuery,
 } from '@/queries/random-facts';
+import { RandomFact } from '@/sanityTypes/RandomFact';
 
 import { sanityClient } from '../../../sanity/lib/client';
-
-import { RandomFact } from '@/types/RandomFact';
 
 const randomFactsApi = async (req: NextApiRequest, res: NextApiResponse) => {
   const trueFacts: RandomFact[] = await sanityClient.fetch(
