+++ src/pages/api/schools.ts
@@ -1,10 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import { schoolsQuery } from '@/queries/schools';
+import { School } from '@/sanityTypes/School';
 
 import { sanityClient } from '../../../sanity/lib/client';
-
-import { School } from '@/types/School';
 
 const schoolsApi = async (req: NextApiRequest, res: NextApiResponse) => {
   const schools: School[] = await sanityClient.fetch(schoolsQuery);
