+++ src/pages/api/short-texts.ts
@@ -1,10 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import { shortTextQuery } from '@/queries/short-texts';
+import { ShortText } from '@/sanityTypes/ShortText';
 
 import { sanityClient } from '../../../sanity/lib/client';
-
-import { ShortText } from '@/types/ShortText';
 
 const shortTextsApi = async (req: NextApiRequest, res: NextApiResponse) => {
   const shortTexts: ShortText[] = await sanityClient.fetch(shortTextQuery);
