+++ src/pages/api/skills.ts
@@ -1,10 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import { skillQuery } from '@/queries/skills';
+import { Skill } from '@/sanityTypes/Skill';
 
 import { sanityClient } from '../../../sanity/lib/client';
-
-import { Skill } from '@/types/Skill';
 
 const skillsApi = async (req: NextApiRequest, res: NextApiResponse) => {
   const skills: Skill[] = await sanityClient.fetch(skillQuery);
