+++ src/pages/api/true-random-facts.ts
@@ -1,10 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 
 import { selectedTrueRandomFactsQuery } from '@/queries/random-facts';
+import { RandomFact } from '@/sanityTypes/RandomFact';
 
 import { sanityClient } from '../../../sanity/lib/client';
-
-import { RandomFact } from '@/types/RandomFact';
 
 const trueFactsApi = async (req: NextApiRequest, res: NextApiResponse) => {
   const trueFacts: RandomFact[] = await sanityClient.fetch(
