+++ src/queries/languages.ts
@@ -1,3 +1,4 @@
+import { gql } from '@apollo/client';
 import { groq } from 'next-sanity';
 
 export const languageQuery = groq`
@@ -7,3 +8,24 @@ export const languageQuery = groq`
   level,
   "flagUrl": flag.asset->url
 }`;
+
+export const languagesQueryQL = gql`
+  query {
+    languages(locale: "en") {
+      data {
+        attributes {
+          name
+          level
+          flag {
+            data {
+              attributes {
+                url
+                alternativeText
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+`;
