+++ src/sanityTypes/HomeContent.ts
