+++ src/sanityTypes/Icon.ts
@@ -1,0 +1,5 @@
+export interface Icon {
+  _id: string;
+  title: string;
+  url: string;
+}
