+++ src/sanityTypes/Language.ts
@@ -1,8 +1,6 @@
-import { Icon } from '@/types/Icon';
-
 export interface Language {
   id: number;
   name: string;
   level: string;
-  flag: Icon;
+  flagUrl: string;
 }
