+++ src/sanityTypes/School.ts
