+++ src/sanityTypes/VideoGame.ts
