--- src/types/Icon2.ts
@@ -1,6 +1,0 @@
-export interface Icon2 {
-  id: string;
-  name: string;
-  url: string;
-  alternativeText?: string;
-}
